backend/requirements.txt
fastapi
uvicorn
pymysql
librouteros
python-dotenv
pysnmp
